Move the suggestion problem to a separate file
